requirements.txt
asgiref==3.8.1
attrs==25.1.0
boto3==1.36.16
botocore==1.36.16
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
django-storages==1.14.4
djangorestframework==3.15.2
drf-spectacular==0.28.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
proto-plus==1.26.0
protobuf==5.29.3
psycopg==3.2.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
